Fixed armors breaking when they shouldn't
Since the plugin may handle the durability of some items, we have to update the event accordingly when we do so. changeDurability method now returns if we hace made any modification to the item. When it returns true, we set damage or repairamount to 0 so we don't break or repair an item. This fixes armors breaking when they shouldn't and items with mending displaying an incorrect damage after being repaired.